Removed unused CSS styles
