+++ web/src/main/scala/net/wiringbits/components/widgets/EditUserForm.scala
@@ -1,7 +1,17 @@
 package net.wiringbits.components.widgets
 
+import com.alexitc.materialui.facade.materialUiCore.createMuiThemeMod.Theme
 import com.alexitc.materialui.facade.materialUiCore.{components => mui, materialUiCoreStrings => muiStrings}
+import com.alexitc.materialui.facade.materialUiStyles.makeStylesMod.StylesHook
+import com.alexitc.materialui.facade.materialUiStyles.mod.makeStyles
+import com.alexitc.materialui.facade.materialUiStyles.withStylesMod.{
+  CSSProperties,
+  StyleRulesCallback,
+  Styles,
+  WithStylesOptions
+}
 import net.wiringbits.api.models.GetCurrentUser
+import net.wiringbits.api.utils.Formatter
 import net.wiringbits.forms.UpdateInfoFormData
 import net.wiringbits.models.User
 import net.wiringbits.ui.components.inputs.{EmailInput, NameInput}
@@ -9,6 +19,7 @@ import net.wiringbits.webapp.utils.slinkyUtils.components.core.ErrorLabel
 import net.wiringbits.webapp.utils.slinkyUtils.components.core.widgets.{CircularLoader, Container}
 import net.wiringbits.webapp.utils.slinkyUtils.forms.StatefulFormData
 import net.wiringbits.{AppContext, AppStrings}
+import org.scalablytyped.runtime.StringDictionary
 import org.scalajs.dom
 import org.scalajs.macrotaskexecutor.MacrotaskExecutor.Implicits.global
 import slinky.core.annotations.react
@@ -20,8 +31,23 @@ import scala.util.{Failure, Success}
 
 @react object EditUserForm {
   case class Props(ctx: AppContext, user: User, response: GetCurrentUser.Response, onSave: () => Unit)
+  private lazy val useStyles: StylesHook[Styles[Theme, Unit, String]] = {
+    val stylesCallback: StyleRulesCallback[Theme, Unit, String] = theme =>
+      StringDictionary(
+        "userEditSummaryView" -> CSSProperties()
+          .setDisplay("flex")
+          .set("& > *", CSSProperties().setMarginRight(16))
+          .set("& > *:last-child", CSSProperties().setMarginRight(0)),
+        "section" -> CSSProperties()
+          .setFlex(1)
+          .setDisplay("flex")
+      )
+
+    makeStyles(stylesCallback, WithStylesOptions())
+  }
 
   val component: FunctionalComponent[Props] = FunctionalComponent[Props] { props =>
+    val classes = useStyles(())
     val (hasChanges, setHasChanges) = Hooks.useState(false)
     val (formData, setFormData) = Hooks.useState(
       StatefulFormData(
@@ -107,6 +133,12 @@ import scala.util.{Failure, Success}
         .`type`(muiStrings.submit)
     }
 
+    val createdAt =
+      Fragment(
+        mui.Typography("Created at").variant(muiStrings.subtitle2),
+        mui.Typography(Formatter.instant(props.response.createdAt))
+      )
+
     form(onSubmit := (handleSubmit(_)))(
       nameInput,
       emailInput,
@@ -116,6 +148,7 @@ import scala.util.{Failure, Success}
           child = ErrorLabel(text)
         )
       },
+      createdAt,
       saveButton
     )
   }
